files/microservices/direct_restapi/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
psycopg2-binary==2.7.5